backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
chromadb>=0.5.0,<0.6.0
beautifulsoup4>=4.12.0
